tools/blip_infer_server/requirements.txt
fastapi
uvicorn
transformers>=4.39
torch
Pillow
accelerate
huggingface_hub

